Refactor to move methods to service
As suggested by @sbearcsiro in CR.
